requirements.txt
fastapi
uvicorn[standard]
transformers
torch
slowapi
python-dotenv

